requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-eventtools==1.0.0
django-filter==2.4.0
django-heroku==0.3.1
docker==4.3.1
Faker==4.14.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jieba==0.39
msgpack==1.0.0
msgpack-python==0.5.6
oauthlib==3.1.0
pbr==3.1.1
postgis==1.0.4
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
twilio==6.46.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
websocket-client==0.57.0
whitenoise==5.2.0
zope.interface==5.2.0
